Fix <li> tag inside the <ul> tag
- The <li> tag inside the homepage and recipe sites has been modified to display as 'inline-block'.
- Each <div> tag inside the recipe sites has applied with appropriate 'background-color' and 'margin'.

The <li> tag changes in the homepage helps for the wider screens.
The <li> tag changes in the recipe sites makes the list aligned properly to the left of the list symbol.
The <div> changes made the site to look more vertically than horizontally as the content was too wide earlier.